Added new options and fixed bugs
Added -c (convert from fasta -> fastq and visa-versa) and -q (if
convert from fasta -> fastq, this option defined what quality
value to use).

Also fixed some command-line argument bugs that would cause a segfault
(like using a non-numeric value for options that want a numeric).

Also fixed a bug where the exclude option was not being used at all,
so the output using it was the same as not using it.